# backend/openapi.json
{
  "components": {
    "schemas": {
      "AcademicCalendar": {
        "properties": {
          "dates": {
            "items": {
              "$ref": "#/components/schemas/CalendarDate"
            },
            "title": "Dates",
            "type": "array"
          },
          "month": {
            "$ref": "#/components/schemas/CalendarMonthMetadata"
          },
          "term": {
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "term",
          "month"
        ],
        "title": "AcademicCalendar",
        "type": "object"
      },
      "AcademicContextResponse": {
        "properties": {
          "administrative_term": {
            "$ref": "#/components/schemas/AcademicTerm"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "is_in_teaching_week": {
            "title": "Is In Teaching Week",
            "type": "boolean"
          },
          "teaching_term": {
            "$ref": "#/components/schemas/AcademicTerm"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          }
        },
        "required": [
          "date",
          "administrative_term",
          "teaching_term",
          "is_in_teaching_week"
        ],
        "title": "AcademicContextResponse",
        "type": "object"
      },
      "AcademicProgress": {
        "properties": {
          "completed_courses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Courses"
          },
          "completed_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Credits"
          },
          "credit_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Score"
          },
          "cutoff_term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cutoff Term"
          },
          "major_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major Rank"
          },
          "major_student_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Major Student Count"
          },
          "remaining_courses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Courses"
          },
          "remaining_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Credits"
          },
          "required_courses": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Courses"
          },
          "required_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Credits"
          }
        },
        "title": "AcademicProgress",
        "type": "object"
      },
      "AcademicProgressCategory": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "completed_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed Credits"
          },
          "converted_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Converted Credits"
          },
          "convertible_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Convertible Credits"
          },
          "course_nature": {
            "default": "",
            "title": "Course Nature",
            "type": "string"
          },
          "course_nature_code": {
            "default": "",
            "title": "Course Nature Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "name_en": {
            "default": "",
            "title": "Name En",
            "type": "string"
          },
          "remaining_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Credits"
          },
          "remark": {
            "default": "",
            "title": "Remark",
            "type": "string"
          },
          "required_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Credits"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "AcademicProgressCategory",
        "type": "object"
      },
      "AcademicProgressCourse": {
        "properties": {
          "college": {
            "default": "",
            "title": "College",
            "type": "string"
          },
          "counts_toward_requirement": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counts Toward Requirement"
          },
          "course_category": {
            "default": "",
            "title": "Course Category",
            "type": "string"
          },
          "course_category_code": {
            "default": "",
            "title": "Course Category Code",
            "type": "string"
          },
          "course_code": {
            "title": "Course Code",
            "type": "string"
          },
          "course_name": {
            "title": "Course Name",
            "type": "string"
          },
          "course_name_en": {
            "default": "",
            "title": "Course Name En",
            "type": "string"
          },
          "course_nature": {
            "default": "",
            "title": "Course Nature",
            "type": "string"
          },
          "course_nature_code": {
            "default": "",
            "title": "Course Nature Code",
            "type": "string"
          },
          "credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hours"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_required": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Required"
          },
          "major_direction": {
            "default": "",
            "title": "Major Direction",
            "type": "string"
          },
          "major_direction_code": {
            "default": "",
            "title": "Major Direction Code",
            "type": "string"
          },
          "module_id": {
            "default": "",
            "title": "Module Id",
            "type": "string"
          },
          "module_name": {
            "default": "",
            "title": "Module Name",
            "type": "string"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          },
          "score": {
            "default": "",
            "title": "Score",
            "type": "string"
          },
          "term": {
            "default": "",
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_code",
          "course_name"
        ],
        "title": "AcademicProgressCourse",
        "type": "object"
      },
      "AcademicProgressCourses": {
        "properties": {
          "categories": {
            "items": {
              "$ref": "#/components/schemas/AcademicProgressCategory"
            },
            "title": "Categories",
            "type": "array"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/AcademicProgressCourse"
            },
            "title": "Courses",
            "type": "array"
          },
          "is_available": {
            "title": "Is Available",
            "type": "boolean"
          }
        },
        "required": [
          "is_available"
        ],
        "title": "AcademicProgressCourses",
        "type": "object"
      },
...
      "AcademicProgressModules": {
        "properties": {
          "is_available": {
            "title": "Is Available",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AcademicProgressModule"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "is_available"
        ],
        "title": "AcademicProgressModules",
        "type": "object"
      },
      "AcademicTerm": {
        "properties": {
          "code": {
            "description": "规范化学年学期代码",
            "title": "Code",
            "type": "string"
          },
          "semester": {
            "description": "学期代码，1/2/3",
            "title": "Semester",
            "type": "string"
          },
          "year": {
            "description": "学年，如 2025-2026",
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "year",
          "semester",
          "code"
        ],
        "title": "AcademicTerm",
        "type": "object"
      },
      "AcademicTermOption": {
        "properties": {
          "code": {
            "description": "规范化学年学期代码",
            "title": "Code",
            "type": "string"
          },
          "is_current": {
            "default": false,
            "title": "Is Current",
            "type": "boolean"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "name_en": {
            "default": "",
            "title": "Name En",
            "type": "string"
          },
          "semester": {
            "description": "学期代码，1/2/3",
            "title": "Semester",
            "type": "string"
          },
          "year": {
            "description": "学年，如 2025-2026",
            "title": "Year",
            "type": "string"
          }
        },
        "required": [
          "year",
          "semester",
          "code"
        ],
        "title": "AcademicTermOption",
        "type": "object"
      },
      "AcademicWarning": {
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "counted_credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counted Credits"
          },
          "earned_courses": {
            "items": {
              "$ref": "#/components/schemas/AcademicWarningCourse"
            },
            "title": "Earned Courses",
            "type": "array"
          },
          "has_warning": {
            "title": "Has Warning",
            "type": "boolean"
          },
          "is_acknowledged": {
            "title": "Is Acknowledged",
            "type": "boolean"
          },
          "is_published": {
            "title": "Is Published",
            "type": "boolean"
          },
          "term": {
            "title": "Term",
            "type": "string"
          },
          "unearned_courses": {
            "items": {
              "$ref": "#/components/schemas/AcademicWarningCourse"
            },
            "title": "Unearned Courses",
            "type": "array"
          }
        },
        "required": [
          "term",
          "has_warning",
          "is_published",
          "is_acknowledged"
        ],
        "title": "AcademicWarning",
        "type": "object"
      },
      "AcademicWarningCourse": {
        "properties": {
          "course_category": {
            "default": "",
            "title": "Course Category",
            "type": "string"
          },
          "course_code": {
            "title": "Course Code",
            "type": "string"
          },
          "course_name": {
            "title": "Course Name",
            "type": "string"
          },
          "course_nature": {
            "default": "",
            "title": "Course Nature",
            "type": "string"
          },
          "credits": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credits"
          },
          "exam_attempt": {
            "default": "",
            "title": "Exam Attempt",
            "type": "string"
          },
          "hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hours"
          },
          "score": {
            "default": "",
            "title": "Score",
            "type": "string"
          },
          "term": {
            "default": "",
            "title": "Term",
            "type": "string"
          }
        },
        "required": [
          "course_code",
          "course_name"
        ],
        "title": "AcademicWarningCourse",
        "type": "object"
      },
      "CalendarDate": {
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "week": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week"
          }
        },
        "required": [
          "date"
        ],
        "title": "CalendarDate",
        "type": "object"
      },
      "CalendarMonthMetadata": {
        "properties": {
          "days_in_month": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days In Month"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "month": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month",
            "type": "integer"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "month"
        ],
        "title": "CalendarMonthMetadata",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "USTB Manager",
      "url": "https://github.com/yourusername/ustb-manager"
    },
    "description": "\n## USTB Manager - 北科大教务管理系统API\n\n这是一个用于访问北京科技大学教务系统的API服务，提供以下功能：\n\n### 🔐 认证模块 (Authentication)\n- **二维码登录**: 使用微信扫码快速登录\n- **短信验证码登录**: 通过手机号接收验证码登录\n- **Cookie登录**: 使用浏览器Cookie直接登录（高级用户）\n- **会话管理**: 自动管理用户会话，支持后端重启后恢复\n\n### 📊 成绩查询模块 (Grades)\n- **成绩列表**: 支持分页、筛选的成绩查询\n- **GPA计算**: 自动计算总GPA和加权GPA\n- **学生信息**: 获取学生基本信息和完整档案\n- **培养方案**: 查看专业培养方案和课程要求\n- **学业进度**: 查询必修课完成情况和学分统计\n\n### 🔒 安全特性\n- HttpOnly Cookie保护\n- Session自动过期和清理\n- Cookie有效性验证\n- 支持CORS跨域访问\n\n### 📝 使用说明\n1. 首先调用认证接口（QR/SMS/Cookie）完成登录\n2. 登录成功后会自动设置session cookie\n3. 后续请求会自动携带cookie进行认证\n4. 可以调用 `/api/auth/status` 检查登录状态\n5. 使用 `/api/auth/logout` 退出登录\n\n### 🌐 环境信息\n- 开发环境: http://localhost:8000\n- API前缀: `/api`\n- 文档地址: `/docs` (Swagger UI) 或 `/redoc` (ReDoc)\n\n### 📞 技术支持\n如有问题，请查看各接口的详细文档说明。\n    ",
    "license": {
      "name": "MIT License"
    },
    "title": "USTB Manager API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/academic/calendar": {
      "get": {
        "operationId": "academic_calendar_api_academic_calendar_get",
        "parameters": [
          {
            "description": "学年学期代码",
            "in": "query",
            "name": "term",
            "required": true,
            "schema": {
              "description": "学年学期代码",
              "pattern": "^\\d{4}-\\d{4}-[123]$",
              "title": "Term",
              "type": "string"
            }
          },
          {
            "description": "月份",
            "in": "query",
            "name": "month",
            "required": true,
            "schema": {
              "description": "月份",
              "maximum": 12,
              "minimum": 1,
              "title": "Month",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicCalendar"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取学期校历",
        "tags": [
          "academic"
        ]
      }
    },
    "/api/academic/context": {
      "get": {
        "operationId": "academic_context_api_academic_context_get",
        "parameters": [
          {
            "description": "查询日期，默认今天",
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "查询日期，默认今天",
              "title": "Date"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取日期感知教学上下文",
        "tags": [
          "academic"
        ]
      }
    },
    "/api/academic/progress": {
      "get": {
        "operationId": "academic_progress_api_academic_progress_get",
        "parameters": [
          {
            "description": "截止学年学期；不传则查询全部学期",
            "in": "query",
            "name": "term",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{4}-[123]$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "截止学年学期；不传则查询全部学期",
              "title": "Term"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgress"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询学业进度汇总",
        "tags": [
          "academic"
        ]
      }
    },
    "/api/academic/progress/courses": {
      "get": {
        "operationId": "academic_progress_courses_api_academic_progress_courses_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgressCourses"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询学业进度课程要求",
        "tags": [
          "academic"
        ]
      }
    },
    "/api/academic/progress/modules": {
      "get": {
        "operationId": "academic_progress_modules_api_academic_progress_modules_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicProgressModules"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询学业进度模块要求",
        "tags": [
          "academic"
        ]
      }
    },
    "/api/academic/terms": {
      "get": {
        "operationId": "academic_terms_api_academic_terms_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AcademicTermOption"
                  },
                  "title": "Response Academic Terms Api Academic Terms Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "获取学年学期列表",
        "tags": [
          "academic"
        ]
      }
    },
    "/api/academic/warnings": {
      "get": {
        "operationId": "academic_warnings_api_academic_warnings_get",
        "parameters": [
          {
            "description": "学年学期；默认使用学业警示模块当前学期",
            "in": "query",
            "name": "term",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{4}-[123]$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "学年学期；默认使用学业警示模块当前学期",
              "title": "Term"
            }
          },
          {
            "in": "cookie",
            "name": "ustb_sid",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ustb Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcademicWarning"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "查询学业警示",
        "tags": [
          "academic"
        ]
      }
    },
...
  },
...
}